Added bridge, finish screen and tweaks
Created a bridge in blender, imported it in Unity and added textures to
it.
Added a small finish screen when the player does three laps.
Fixed a bug with the camera manager: there was an issue with the helper
points rotating with the formula. It should now be fixed
Small tweaks of map and code