Tweaks and terrain update
Added grass, flowers and details to the grass texture.
Tweaked the terrain so it looks better thanks to the new brushes.
Changed the height of arrow2 so the player can see it from afar.
Added background music (BGM) and all the logic behind it (Like slider to
mute the music).
Changed the look of MainMenu scene as well, to accomodate the new
details.